Fix Spark quoted columns (#6)
If a Spark column is selected with a quote, it's currently incorrectly
translated to a string.

e.g.

```
select `column` from table
```

becomes

```
select 'column' from table
```

This PR fixes that. The error was that for Spark queries, that we'd
already transpiled to Trino, we were parsing them as Spark.