Don't replace Spark's timestamp with from_unixtime anymore (#24)
We added a fix to SQLGlot upstream for Spark's timestamp function
(https://github.com/tobymao/sqlglot/pull/1454). The function is just
shorthand for a cast, so now it's translated to a cast. It's still not
strictly correct Trino, since the output type is slightly different, so
will need to fix that.