Translate '0x..' strings into hex strings inside DuneSQL (#26)
In DuneSQL, we want to work with hex values as varbinary values. We
read/output these as `0xdeadbeef`.

Currently we capture these with a regex and replace them in
`remove_quotes_around_0x_strings`:

https://github.com/duneanalytics/harmonizer/blob/2281f40774b0ccecb121dac5a01f6999b35fd70c/dune/harmonizer/custom_transforms.py#L350-L356

This PR replaces that transform with a preprocessing step in the
generator step of the DuneSQL dialect.