Don't remove leading zeros from hex strings (#45)
As described in #41, SQLGlot currently removes leading zeros from hex
strings at parse time. This is bad for us since in Trino `0x00` and
`0x0000` are not the same, nor are `0xdeafbeef` and `0x00deadbfeef`.
Many Dune users use `where col = <hex>` filters, so this can cause a lot
of incorrect results if queries are migrated with this bug.

This PR is a workaround to that. We use the original query and the final
query and stitch back the hex strings. It seems to work correctly,
assuming the new query hasn't been totally switched around and there are
many duplicates.