Add optimize() for transforming DuneSQL expressions using schema (#57)
This PR adds an `optimize()` function. SQLGlot has optimization
machinery which can use table schemas to do transformations on the AST.
It requires that a query is fully qualified and annotated with types. We
don't use SQLGlot's main `optimizer.optimize` yet (we may want to), but
we use helpers from that library.
We add one actual optimization function which involves attempting to
force types being equal in an `EQ` node, i.e. `where 1 = '1'`.
We don't use any of this in the main `translate_` functions yet.