Add alias to simple casts if query is Postgres (#59)
In Postgres, `cast(x) as type` and `x::type` has alias `x`, whereas in
Trino it does not. We should support this simple case, as some users
have come to inadvertently rely on it. It's not the case in Spark and
it's not the case for non-column expressions. Since it's only the case
in Postgres, we need to add it outside the DuneSQL dialect.

Fixes #55