Translate Spark's null safe indexing into null safe element_at function (#77)
Fixes #75.

A lot of Dune queries rely on Spark's indexing operation being null
safe, i.e. it returns `null` if the array isn't long enough, instead of
failing at execution, which is what Trino does. A lot of Solana queries,
where there's a lot of working with columns of arrays, and columns of
arrays of structs, rely on this feature of Spark.

To better accommodate it, we change an indexing operation into Trino's
null safe
[`element_at`](https://trino.io/docs/current/functions/array.html#array-functions)
function. Further, we increment the index by 1, since Spark is 0-indexed
while Trino is 1-indexed.